+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Fixing a bug with `clustering/k-means` where a vector could end up in several clusters.
 * Dropping the internal `regex/classes` namespace.
 * Dropping the `hasher` option of the `ngrams` functions.
+* Dropping the `set-functions` dependency.
 
 ## 0.11.0